requirements.txt
python-telegram-bot==20.2
python-telegram-bot[all]==20.2
python-telegram-bot[bot]==20.2
